seperate the file
